requirements.txt
alembic==1.13.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
flasgger==0.9.7.1
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
Mako==1.3.0
MarkupSafe==2.1.3
mistune==3.0.2
numpy==1.26.2
packaging==23.2
pandas==2.1.4
psycopg2==2.9.9
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-logstash==0.4.8
pytz==2023.3.post1
pyxirr==0.10.1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.15.2
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
